Made some changes to remove the compile warnings...
on the comparison between un-uniform data types

(cherry picked from commit e0705e2a87726c6d8a04eae624d4f032c5d596ce)

Conflicts:
	sw/source/filter/ww8/wrtw8esh.cxx
	sw/source/filter/ww8/wrtww8.cxx
	sw/source/filter/ww8/ww8par.cxx
	sw/source/filter/ww8/ww8par.hxx
	sw/source/filter/ww8/ww8par3.cxx

+++ sw/source/filter/ww8/wrtww8.cxx
@@ -305,7 +305,7 @@ void WW8_WrtBookmarks::MoveFieldMarks(WW8_CP nFrom, WW8_CP nTo)
     {
         if (aItr->second)
         {
-            if (aItr->second->first == nFrom)
+            if (aItr->second->first == (long)nFrom)
             {
                 aItr->second->second.first = true;
                 aItr->second->first = nTo;
@@ -1407,8 +1407,8 @@ int MSWordExportBase::CollectGrfsOfBullets()
 
     if ( pDoc )
     {
-        int nCountRule = pDoc->GetNumRuleTbl().size();
-        for (int n = 0; n < nCountRule; ++n)
+        size_t nCountRule = pDoc->GetNumRuleTbl().size();
+        for (size_t n = 0; n < nCountRule; ++n)
         {
             const SwNumRule &rRule = *( pDoc->GetNumRuleTbl().at(n) );
             sal_uInt16 nLevels = rRule.IsContinusNum() ? 1 : 9;
@@ -1423,7 +1423,7 @@ int MSWordExportBase::CollectGrfsOfBullets()
                 if ( pGraf )
                 {
                     bool bHas = false;
-                    for (unsigned i = 0; i < m_vecBulletPic.size(); ++i)
+                    for (size_t i = 0; i < m_vecBulletPic.size(); ++i)
                     {
                         if (m_vecBulletPic[i]->GetChecksum() == pGraf->GetChecksum())
                         {
@@ -1522,7 +1522,7 @@ int MSWordExportBase::GetGrfIndex(const SvxBrushItem& rBrush)
     int nIndex = -1;
     if ( rBrush.GetGraphic() )
     {
-        for (unsigned i = 0; i < m_vecBulletPic.size(); ++i)
+        for (size_t i = 0; i < m_vecBulletPic.size(); ++i)
         {
             if (m_vecBulletPic[i]->GetChecksum() == rBrush.GetGraphic()->GetChecksum())
             {
